Update patch set 2
Patch Set 2:

Hi Darius, 
I took a full pass on the code. I think there are two things to discuss here.
static storage of the variable in the Expression class and testing of these classes.

If this is done on a separate email thread, we are happy to do that too.

Patch-set: 2
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>