Update patch set 2
Patch Set 2:

Hi Darius,

Keir and I talked about the expression_data a bit more. And here is a small modification which should allow us to get rid of it.

expression_data is doing three things.

1. It is generating new unique ids.
2. Managing the life time of expressions.
3. Collecting all the nodes in the expression tree into a single place.

But what if, ExpressionRef instead of storing an index into expression_storage, stored a shared_ptr<Expression>?

So the ID becomes the pointer allocated by the memory allocator -- which is guaranteed to be unique and the shared_ptr will make sure the lifetime issues are handled correctly. 

This would not give you a full list of expressions at construction, but that is easy to generate posthoc by traversing the tree after we "evaluate" the functor.

Sameer

Patch-set: 2
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>