Update patch set 2
Patch Set 2:

Hi Darius,
My comments are inline.

> > also once expression_storage goes away, it makes more sense to move all the static methods in the Expression class - MakeFoo, to be static methods of whatever ExpressionRef ends up being called.
> > 
> > Talking more to Keir, maybe ExpressionRef should be Expression and Expression should be
> > ExpressionData. This makes even more sense, once we decide to move away from ids to 
> > shared_ptrs.
> 
> Hello Sameer,
> 
> thanks for the review. I like the idea of using shared pointers in expressionsRef. A nice 
> feature would be that unused variables delete them self because they are not referenced by 
> anyone. But thinking a bit further, I see a problem when we get to conditional blocks. 
> 
> How does an expression know in which block it was executed?
> How do we handle nested/sequential if/else?
> 
> It feels like that we need at least one global state variable which holds the current block. > Most likely though a stack<conditionalBlock> or something similar is required. 
> ===============
> 
> Using a global expression vector, conditionals are very easy to implemented. If we pass by the CERES_IF macro, an _IF_ expression is inserted into the vector and all later expressions (until we hit an CERES_ENDIF) are inside this block.

Very interesting catch. We did not realize that the order of expressions in expression_storage was meaningful and was being used to keep track of conditionals.

> To make the expression vector better and conforming your coding standards, I would suggest the following changes:
> 
> 1. Move the variable to expression.cc so the symbol is not visible.
> 2. Make it a pointer (trivial constructor).
> 3. Add a "StartCodeGeneration" and "EndCodeGeneration" method.

This is basically the Singleton pattern. It still makes testing icky but I think this is strictly better than having the vector be a static member of the Expression object.

I will think a bit more about if its possible to get rid of the singleton, but I do not want to block you. This change itself will lead to a considerable amount of simplifcation of the current code, so lets go ahead with it.

Patch-set: 2
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>