Update patch set 4
Patch Set 4:

Hello,

this patch is now split up into three files:

1. Expression.h - Holds the expression data (types/parameters and so on)
2. ExpressionRef.h - The type used for arithmetic 
3. ExpressionTree.h - Contains the expression list which is generated during execution.

I want to discuss the following two points.

- Change the naming?

I chose the current names, because they match to the implementation. "Expression" contains all the data for one line of code. "ExpressionRef" only contains the index which references an Expression.

- Move the static MakeExpression* functions to ExpressionRef.h?

It probably makes more sense to have the MakeExpression* function in expressionref, because they take ExpressionRefs as parameters and return an ExpressionRef.

Patch-set: 4