Update patch set 4
Patch Set 4:

(35 comments)

Darius, more questions and comments for you. I restricted to myself to the header files, since thats the meat of the discussion for now IMO. the implementation will follow from the headers in a straightforward manner.

To answer the questions you raised

>Change the naming?
>I chose the current names, because they match to the implementation. "Expression" contains >all the data for one line of code. "ExpressionRef" only contains the index which references >an Expression.

As I mention in my comments, ExpressionTree should be called ExpressionGraph, but I suspect that you can get rid of the class entirely, also I think we can simplify the relationship between Expression and ExpressionRef. 

>Move the static MakeExpression* functions to ExpressionRef.h?
>It probably makes more sense to have the MakeExpression* function in expressionref, because >they take ExpressionRefs as parameters and return an ExpressionRef.

I think this depends on what happens to the relationship between Expression and ExpressionRef, why don't we resolve that and the need for an ExpressionGraph/Tree class (instead of just having a vector) and then make a decision about this?

Patch-set: 4
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>