Update patch set 6
Patch Set 6:

(11 comments)

> Patch Set 5:
> 
> I think the next set of changes to the API will likely be driven by the needs of testing and how we can cleanly test expressions, expressionref and expressiontree/graph.

Fixed minor issues + renamed Tree->Graph.

I would suggest that we add test for all methods except Expression::CreateXXX and the ExpressionRef operators in this patch. It's much easier to test the operators and Create functions when we have a simple code generator. Then we can just write micro cost functors generate code for them and compare the results to the reference.

Patch-set: 6