Update patch set 11
Patch Set 11:

(6 comments)

A couple of more minor comments, otherwise this is good to go as far as the API and code goes.
The testing needs more work. I do not think depending on the code gen is a great idea, because I expect the code gen to change over time.

I think we should test small expression trees on their own. It does not have to be in this CL, as it is large enough already, but I think we should land the conditional stuff next and then test the expression infrastructure thoroughly before moving onto code generation.

Patch-set: 11
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>