Update patch set 1
Patch Set 1:

(6 comments)

Darius, I took a quick look at this, and handling conditionals seems to make the expression api considerably more complicated and hard to understand. Also the patch has very little testing (I really do not want to depend on the code generator for that). 

Could you add more example code and documentation for why the multi-assignment semantics are needed? and where the ssa form is breaking. 

if the comments do not belong in the code, the rationale for the design should be thoroughly documented in the commit log.

Patch-set: 1
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>