Update patch set 6
Patch Set 6:

(18 comments)

Hi Darius,

With your expanded comment I am starting to understand where this is going. I realize now that the ExpressionGraph is not in fact a graph representable as an unordered sequence of (id, op, [args]*), but instead is an ordered list of instructions, which may or may not depend on the result of other, previous, instructions; that may or may not be in SSA form.

Now we are getting closer to the "big compiler" level of complexity with control flow. However, this code appears to try and avoid the traditional data structures like basic blocks, through the IF/ELSE/ENDIF expression types added in this patch.

How are you planning on handling control flow in the codgen phase? Are you going to build code to convert to full SSA form? Do you expect to add basic block handling?

Keir

Patch-set: 6
Reviewer: Gerrit User 5001 <5001@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>