Update patch set 7
Patch Set 7:

(15 comments)

Hi Keir.

> With your expanded comment I am starting to understand where this is going. I realize now that the ExpressionGraph is not in fact a graph representable as an unordered sequence of (id, op, [args]*), but instead is an ordered list of instructions, which may or may not depend on the result of other, previous, instructions; that may or may not be in SSA form.

Yes that is true.

> Now we are getting closer to the "big compiler" level of complexity with control flow. However, this code appears to try and avoid the traditional data structures like basic blocks, through the IF/ELSE/ENDIF expression types added in this patch.

I think that is kind of the beauty of this approach. Since we have an actual compiler before and after our code optimizations, we can get rid of many complicated traditional data structures such as ASTs and CFGs. But yeah, we will need block/scope handling during optimization (not during code generation though).

> How are you planning on handling control flow in the codgen phase? 

The codegen phase is a looooot simpler than you might expect. We don't have to do any control flow handling in there. It will basically look like this:

for(e : graph.expressions_){
  ExpressionToString(e);
}


// ...

ExpressionToString(e){
  switch(e.type){
    case IF:
      return "if("+e.id+"){";
    case ELSE:
      return "}else{";
    case PLUS: 
       return e.lhs + "=" + e.params[0] + "+" + e.params[1];
    // ...
   }
}

This works, because we have the first compiler phase that ensures a valid list of expressions in the correct order and with correct arguments. Of course, since IF/ELSE are now macros, we can count the number to check if each IF was terminated by an endif.

> Are you going to build code to convert to full SSA form? 

No, most likely not. Assignments to the same variable from different if-branches are very complicated to convert into SSA form. This either requires a completely different data structure (CFGs) or a flattening of the branches into PHI-functions.

> Do you expect to add basic block handling?

Yes, we need block/scope handling, but only during optimization.


Darius

Patch-set: 7