Update patch set 7
Patch Set 7:

(1 comment)

> Patch Set 6:
> 
> (2 comments)
> 
> Keir I'd like you to look at the change to Jet

I like having a non-explicit (double) constructor for ExpressionRef more than the two Jet constructors of the current patch.

I don't see a problem where it can break. ExpressionRef is somewhat a wrapper around a double so a direct construction is valid. Constructing a Jet from a double on the other hand does a lot more.

Patch-set: 7