Update patch set 10
Patch Set 10:

Hey.

This issue is actually more than just me forgetting to run the test case. In fact this test passes on clang, but fails on gcc.

What causes this behaviour is the following line in Jet::Operator*:

  return Jet<T, N>(f.a * g.a, f.a * g.v + f.v * g.a);

On clang it is evaluated left to right, but on gcc the order is reversed. Of course, semantically both are correct, but since this test expects a predefined order, only one can succeed. 

https://en.cppreference.com/w/cpp/language/eval_order

The only quick fix I see is removing the failing test.

Patch-set: 10