Update patch set 24
Patch Set 24:

(1 comment)

the addition of void is a bit of a surprise to me, but lets get this in, there is also the broken test to fix.

Before we start doing the optimizer, I'd like to get the public facing API as to how the user will actually use this checked in next, which will likely inform the re-organization of the code that Keir talked about too. 

once that is in place, we should add benchmarks for a number of cost functions, and then start working on the optimizer.

Patch-set: 24
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>