Autodiff Codegen Part 4: Public API
The public API of the autodiff codegen system consists of the two
classes AutoDiffCodeGenerator and AutoDiffCodeGenCostFunction.

AutoDiffCodeGenCostFunction is an interface class similar to
SizeCostFunction except that it doesn't derive from CostFunction.

AutoDiffCodeGenerator takes as template argument a cost functor
type which dervies from AutoDiffCodeGenCostFunction and generates
the corresponding C++ for the residual and derivative. This class
mainly serves as a wrapper for the different codegen modules.

Change-Id: I8ee974199219805d54eed5e07c0e8d1394940779