Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (12 comments)
> 
> I currently see 3 ways to solve the problem of some jacobians being null
> 
> (1) Generate if expressions around the output assignments to the jacobians. This requires an additional ExpressionType which can handle "jacobian[0]==nullptr".
> 
> (2) Compute all jacobians into a temp array and let the combined Evaluate function copy if necessary.
> 
> (3) Generate code for all 2^n possible combinations. Then use a index-lookup to a function pointer in the combined function.

(3) I do not think is a great solution and I am afraid I do not understand the advantage/disadvantage of (1).

So here is some thinking out aloud, why not first compute everything, assuming that all jacobians are to be computed and then post process the assignment to jacobian array code, the only thing that needs to be done is something like

if (jacobians[0] != nullptr) {
  do assignments for jacobians[0]; // this code would have been generated anwyays.
}

but this assumes that the assignment code sits at the bottom and can easily be identified but as I write this, this is basically (2). 

so I guess my recommendation is (2).

Patch-set: 7
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>