Update patch set 7
Patch Set 7:

> (3) I do not think is a great solution and I am afraid I do not understand the advantage/disadvantage of (1).
> 
> So here is some thinking out aloud, why not first compute everything, assuming that all jacobians are to be computed and then post process the assignment to jacobian array code, the only thing that needs to be done is something like
> 
> if (jacobians[0] != nullptr) {
>   do assignments for jacobians[0]; // this code would have been generated anwyays.
> }
> 
> but this assumes that the assignment code sits at the bottom and can easily be identified but as I write this, this is basically (2). 
> 
> so I guess my recommendation is (2).

Personally I like (3), but it doesn't scale very well.

The problem with (1) is that I do not want to make any assumptions about the optimizer yet. New expression may be created or the order may be changed. That means we cannot group the jacobian assignments anymore and have to write them one by one like this:

if(jacobians[0] != nullptr)
  jacobians[0][0] = v_3;
if(jacobians[0] != nullptr)
  jacobians[0][1] = v_3;
if(jacobians[0] != nullptr)
  jacobians[0][2] = v_3;
...

My initial idea was to add if-expressions BEFORE the optimizer, but that requires a new ExpressionType. Adding them before optimization should keep the grouping and enable even further optimizations. For example, moving expressions inside scopes if they are not used by other scopes.

(2) is most likely the best solution. We don't have to forget that there is another compiler trying to optimize the code before execution. If this compiler inlines EvaluateResidualsAndJacobians into Evaluate then all the unnecessary assignments can be removed.

Patch-set: 7