Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)

I don't really see the problem with the Ternary function how it is now. Converting all ?-operators to if/else can't be the solution here. We can also make it to a macro which just outputs the ?-operator outside of CODEGEN mode, but i thought a function would be nicer here.

Patch-set: 6