Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (2 comments)

AddExpressionToGraph is currently in expression_ref because otherwise we would get a cyclic dependency ExpresionRef<->ExpressionGraph.

The other solution would be to add a function AddExpressionToGraph to ExpressionGraph which only returns an ExpressionId instead of ExpressionRef. But that would make ExpressionRef harder to read because the ExpressionId has to be converted explicitly to ExpressionRef in each operation.

I'm not sure what the best solution is here.

Patch-set: 5