Update patch set 1
Patch Set 1:

NeroBurner (I do not know your real name so my apologies), thanks for the awesome work on simplifying Ceres's cmake infra. Please add Alex & Me to all of these patches as a reviewer so that he gets notified when you upload a change.Otherwise I will have to be looking at gerrit to know that you have a patch waiting to be reviewed.

Patch-set: 1