Update patch set 11
Patch Set 11:

with regards to the plan

I think you need to add the logic to deal with constant/null jacobians before you do any of the rest.
also it maybe a good idea to file github issues for each of these, and that way we can discuss there and track progress instead of doing this over email.


My current plan is:

2. CMake integration
3. Tests for rotation.h (compare result of generated code with autodiff)
4. Benchmarks for bundleadjustment without any optimiations
5. Optimization

Patch-set: 11