Update patch set 4
Patch Set 4:

(8 comments)

The thing that I don't like right now is that the input functor and the generated cost function are in two different classes. That means, when we add a new local variable it has to be added to both types.

I going to solve this by adding a new CostFunctor interface which changes behavior during code generation.

Patch-set: 4