Update patch set 7
Patch Set 7:

(6 comments)

This looks good, however we will have to install FindSuiteSparse.cmake with Ceres (at least when it is statically compiled) as in that case all dependencies, even ones declared PRIVATE are actually (and necessarily) treated as PUBLIC by CMake.  As it stands, that would be a problem because the Ceres target will reference the imported targets, not the original libraries and the imported targets are not re-exported when Ceres is exported and so the result is that Ceres would be unusable as its target would depend on targets that do not exist and would fail when a user tried to compile their code that used Ceres.

Patch-set: 7
Reviewer: Gerrit User 5155 <5155@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>