Update patch set 9
Patch Set 9:

(21 comments)

This is a nice idea, and the basis for it is there but I think we need to refine it to make it portable and usable outside of Ceres in a client project.

Darius a good way of testing this would be to create a trivial helloworld client CMake project and try to call this from there, that would highlight the bulk of the issues raised here.

Patch-set: 9
Reviewer: Gerrit User 5155 <5155@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>