Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> (2 comments)
> 
> Minor questions, however I guess my larger question is should this have a string accessor that updated the total (and then avoiding the use of the destructor for this purpose) rather than defining that it only logs to VLOG(3)?

So the reason for the destructor is that, this is technically a "ScopedEventLogger", where we just stick this sucker in, and regardless of how we exit the function especially with complicated exit conditions, we always log. Otherwise you will have to grab the accessor everytime you exist.

Patch-set: 2