Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> OK, so reading this again I wonder whether we shouldn't propagate the IsConstant()/IsSetConstantByUser() through to the Problem API in light of this change as it seems like the current API isn't really reflecting the state of the problem after this change is merged.

Here is my thinking around this.

It comes down to what the value returned by IsParameterBlockConstant should be? when a local parameterization with a tangent space of size zero is associated with a parameter block.

IMO A parameter block being marked constant by the user and it being effectively constant because its manifold has collapsed to a point are equivalent but not the same thing.

IsParameterBlockConstant should only return true if the user has explicitly marked it constant. Doing it based on the dimension of the tangent space makes its semantics more complicated to explain.


WDYT?

Patch-set: 6