Update patch set 1
Patch Set 1:

In this CL a dynamic only version of the dimension is implemented to be consistent with HomogenousVectorParameterization. This means the constructor takes the dimension at runtime. I ran a benchmark of this version (dynamic), a static only, where the dimension is specified as a template and a mixed one which can handle static and dynamic dimensions. IMO in a lot of cases, the dimension is known at compile time (e.g. optimize a line in 3D):
  ------------------------------------------------------------------------------
  Benchmark                                       Time           CPU Iterations
  ------------------------------------------------------------------------------
  BM_LineParameterizationStatic                  44 ns         44 ns   15740220
  BM_LineParameterizationDynamic                168 ns        168 ns    4181117
  BM_LineParameterizationMixedStatic             49 ns         49 ns   13654458
  BM_LineParameterizationMixedDynamic           165 ns        164 ns    4195890
  BM_LineParameterizationJacStatic               25 ns         25 ns   27286711
  BM_LineParameterizationJacDynamic              87 ns         87 ns    7979200
  BM_LineParameterizationJacMixedStatic          29 ns         29 ns   24075668
  BM_LineParameterizationJacMixedDynamic         81 ns         81 ns    8385686

The static version is about 4x faster for the Plus operator and 3x faster for the Jacobian evaluation. The mixed version is almost as fast as the static one if the dimension is specified at compile time and as fast as the dynamic one in case it is specified at runtime.

I am not sure, if it is better to be consistent with the HomogenousVectorParameterization or use a templated faster version. What do you think?

Also the HomogenousVectorParameterization could be rewritten to support both, static and dynamic (the mixed case), but I am not exactly sure how to maintain the API in C++11.

Patch-set: 1
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>
Reviewer: Gerrit User 5005 <5005@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>