Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks Johannes, what do you think about a slightly more general affine subspace parameterization instead which will reduce to lines for the one dimensional case?

What do you exactly mean by affine subspace parameterization? E.g. in 3D space a plane or line? If so, the question is how to parameterize the plane or line and how a common local parameterization look like? A line in 3D has 4 DOF and a plane 3 DOF. A line is conveniently parameterized as two points or a point and a direction and a plane can be modeled with a normal vector and a distance (Hesse normal form). I am not sure, how one local parameterization can be used in both cases.

I am not sure, if my thinking goes in a totally wrong direction. Do you have some more examples, details or use cases?

Patch-set: 1