Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Thanks Johannes, what do you think about a slightly more general affine subspace parameterization instead which will reduce to lines for the one dimensional case?
> 
> What do you exactly mean by affine subspace parameterization? E.g. in 3D space a plane or line? If so, the question is how to parameterize the plane or line and how a common local parameterization look like? A line in 3D has 4 DOF and a plane 3 DOF. A line is conveniently parameterized as two points or a point and a direction and a plane can be modeled with a normal vector and a distance (Hesse normal form). I am not sure, how one local parameterization can be used in both cases.
> 
> I am not sure, if my thinking goes in a totally wrong direction. Do you have some more examples, details or use cases?

Let "c" be a point in R^n.

And let M be an n x k orthonormal matrix. Then 

The affine subspace is given by points of the form . c + My.

where y is a k dimensional vector.

so it is a k dimensional space "centered" at "c".

A line is a special case for k = 1.

A plane can be parameterized with k = 2, by having M have two columns which form the basis of the points in the plane.

One design point worth considering here is,  whether we demand that the user provide us with M which is orthonormal, or do we construct an orthonormal basis out of it via a QR factorization? If we do demand that M is orthonormal and the user provides us with a matrix which is not orthonormal then what do we do?

Patch-set: 1
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>