Fix windows MSVC build.
This CL fixes the build with newer MSVC versions. It adds
the cmake option MSVC_ENABLE_EXTENDED_ALIGNED_STORAGE
to allow to use proper over-aligned types which is needed by Eigen.
The default is to enable it, which is ABI incompatible with older
MSVC versions. It also adds the flag /bigobj otherwise the build of
the unit test fails.

Change-Id: I63f08cd6c0a8db8c9931289b909b4deafd75b039