Update patch set 3
Patch Set 3:

I thought about the whole problem in more detail and think, it is better to remove std::aligned_storage from FixedArray (which is done by the Abseil team in absl::FixedArray). This fixes the build for older MSVC versions as std::aligned_storage is non-conformant and prevents the use of the extended aligned storage macro, which could affect user code.

Patch-set: 3