Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > Thanks Johannes, what do you think about a slightly more general affine subspace parameterization instead which will reduce to lines for the one dimensional case?
> > > > 
> > > > What do you exactly mean by affine subspace parameterization? E.g. in 3D space a plane or line? If so, the question is how to parameterize the plane or line and how a common local parameterization look like? A line in 3D has 4 DOF and a plane 3 DOF. A line is conveniently parameterized as two points or a point and a direction and a plane can be modeled with a normal vector and a distance (Hesse normal form). I am not sure, how one local parameterization can be used in both cases.
> > > > 
> > > > I am not sure, if my thinking goes in a totally wrong direction. Do you have some more examples, details or use cases?
> > > 
> > > Let "c" be a point in R^n.
> > > 
> > > And let M be an n x k orthonormal matrix. Then 
> > > 
> > > The affine subspace is given by points of the form . c + My.
> > > 
> > > where y is a k dimensional vector.
> > > 
> > > so it is a k dimensional space "centered" at "c".
> > > 
> > > A line is a special case for k = 1.
> > > 
> > > A plane can be parameterized with k = 2, by having M have two columns which form the basis of the points in the plane.
> > > 
> > > One design point worth considering here is,  whether we demand that the user provide us with M which is orthonormal, or do we construct an orthonormal basis out of it via a QR factorization? If we do demand that M is orthonormal and the user provides us with a matrix which is not orthonormal then what do we do?
> > 
> > First, I wish you a happy new year!
> > 
> > Thanks for the details. I am still thinking about how to formulate a box plus operator for orthogonal matrices (M). The k=1 case is in this CL. But I don't know yet how this can be generalized for k > 1.
> 
> If the point X is already of the form x = c + M y, for some y.
> 
> and you get a new point delta, then the plus operator is 
> 
> x_plus_delta = x + M delta = c + M (y + delta) 
> 
> 
> which means that you do not even need to know what c is, you only need to know the matrix M.

Sorry, but now I am confused. Let's consider the following problem: find the best 3D line which is closest to n 3D points (only a simplified example which could be solved directly without formulating a NLS problem). Optimizing just y without M means, that the line direction is constant. The parameters are the 1D position on the line (y) and the 3D origin point c. So the local parameterization with constant M is not usable. Also the local parameterization for c is missing as it only has 2 DOF. That is what this CL provides (a local parameterization for c and M if k = 1).

I think, it would be nice to generalize this CL for k>1 but I am not sure if this can be solved. E.g. for n = 3 and k = 3 this means, that M could be a rotation matrix and an update of a rotation matrix with just 3 parameters is not trivial (I would probably use an exponential map and use M_R * exp(skew(y)) as the box plus operator). But this also means, that the number of optimization parameters is 9 and the DOF are 3. I don't know if this has performance penalties.

Patch-set: 1