Update patch set 26
Patch Set 26:

Thanks for your patience with this Darius. I am going to go ahead and merge this. It is likely that we are going to iterate on the cmake integration and some naming stuff some more, but I'd rather kick the tires on this in another build system - bazel at google to see what it feels like than just iterate on what is already a very long code review.

Patch-set: 26
Reviewer: Gerrit User 5002 <5002@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>