Update patch set 1
Patch Set 1:

> This seems reasonable based purely on my reading of the docs for WINDOWS_EXPORT_ALL_SYMBOLS - although I'm not an MSVC user, so my only questions are around what this does to the size of the resulting .dll - is it significantly increased by this or is there any other downside to this?

Hm. Good point. Looking at our build, ceres.dll when from 7.5 MB to 24.7 MB with this change. Perhaps the better solution is to keep the existing check for BUILD_TESTING and BUILD_SHARED_LIBS but then add an additional check for WINDOWS_EXPORT_ALL_SYMBOLS is enabled before forcing BUILD_TESTING to OFF. (and update the message to suggest to the user that they build with WINDOWS_EXPORT_ALL_SYMBOLS=ON if they want unit tests)

Patch-set: 1