Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (4 comments)
> 
> sorry for the delay in replying to you, life got a little crazy but I am back now. lets get this in.

Thanks for reviewing. I will look into it but I do have a question upfront: In the current implementation the space dimension of the line (e.g. 2D or 3D) is specified as a runtime variable. If this parameter is known at compile time it would be roughly 4x faster. In a comment above I added the timings (comment on Dec 29 2:31 PM):
  ------------------------------------------------------------------------------
  Benchmark                                       Time           CPU Iterations
  ------------------------------------------------------------------------------
  BM_LineParameterizationStatic                  44 ns         44 ns   15740220
  BM_LineParameterizationDynamic                168 ns        168 ns    4181117
  BM_LineParameterizationMixedStatic             49 ns         49 ns   13654458
  BM_LineParameterizationMixedDynamic           165 ns        164 ns    4195890
  BM_LineParameterizationJacStatic               25 ns         25 ns   27286711
  BM_LineParameterizationJacDynamic              87 ns         87 ns    7979200
  BM_LineParameterizationJacMixedStatic          29 ns         29 ns   24075668
  BM_LineParameterizationJacMixedDynamic         81 ns         81 ns    8385686

The current implementation would be BM_LineParameterizationDynamic. For the BM_LineParameterizationStatic the space dimension needs to be known at compile time. For the BM_LineParameterizationMixed* case the dimension can either be fixed or Eigen::Dynamic, so it supports both. The test is done for a three dimensional line.

To implement the mixed case the class
  class CERES_EXPORT LineParameterization : public LocalParameterization
would change to
  template <int Dimension>
  class CERES_EXPORT LineParameterization : public LocalParameterization
where Dimension can be Eigen::Dynamic. What do you think?

Patch-set: 1