Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> > 
> > One more thing that I thought of.  Is there a unit test you can add to ensure the loss functions do have the property of 1/2 w(0) = 1?  This way if people add a loss function in the future then the testing can catch it.
> 
> I just changed the comment in the header file for the Tukey loss as suggested.

Sorry for the slow response.  Thanks for updating the header. I re-looked at the test and yes I think it would be good to check that the value equals 1 for the other loss functions.  I went back and forth on this but the reason why I think it is good is because if someone adds a new loss function they will most likely copy the loss function test from a previous loss function and update it.  Even though there is no automated way to ensure this will be included most likely it will.

Patch-set: 7
Reviewer: Gerrit User 5435 <5435@cdcfcfbe-2044-3b1f-b5d4-da77ad542329>