Autodiff with Static Arrays
Use std::array to store the Jets for functors with few parameters
or residuals. This allows the compiler to further optimize the
autodiff code. After this change, Jet-Autodiff (almost) reaches
CodeGen performance for small functors (~less than 25 params).
This is around 2x-5x faster than the previous implementation.

This patch also adds kNumResiduals as a template argument to
AutoDifferentiate, so that the output can be stored on the stack
and the Take1stOrder functions can be fully unrolled.

Change-Id: I3604c5673f5a34f88b24ad404b6f1a885e53f498