Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (3 comments)
> 
> minor comments, lets wait on Keir to take a look and for me to test this inside google later today.

Sounds good.

Yes these numbers are very large. In difference to the FixedArray before, we do not allocate that much for every functor. A small functor with 3 parameters and 1 residual only allocates space for 4 Jets on the stack. So having a large number here doesn't hurt, but yeah 50 was probably a bit exaggerated.

Patch-set: 5